run-testsの通常Editor assembly化
RunTests 本体の asmdef から TestAssemblies 扱いを外し、consumer project でも tool registry に登録されるようにする。Unity Test Runner API 依存は Test Framework 有効時だけ compile される adapter assembly に分離し、Test Framework なしの project では従来通り unsupported response を返す。
