+++ Assets/Editor/uLoopMCP.Dev.asmdef
@@ -17,6 +17,7 @@
         "GUID:ee36dcf19655f43c08e0a7009ca7a624",
         "GUID:54dfc5fd447794df2aa683025462d668",
         "GUID:9ed21973545c44e28af522b049dec47d",
+        "UnityCLILoop.FirstPartyTools.RunTests.TestFramework.Editor",
         "GUID:a0bdbd2c5705643fbb9aef9fac8fd46a",
         "GUID:fc3fd32eddbee40e39c2d76dc184957b",
         "GUID:c956a21f824994ef087b6de566690b3d",
