+++ Assets/Tests/Editor/uLoopMCP.Tests.Editor.asmdef
@@ -25,6 +25,7 @@
         "GUID:dc412b2cd951d452b8dcec99c3adeb19",
         "GUID:3b7aa330d87dc452293296dd5830859b",
         "GUID:9ed21973545c44e28af522b049dec47d",
+        "UnityCLILoop.FirstPartyTools.RunTests.TestFramework.Editor",
         "GUID:a0bdbd2c5705643fbb9aef9fac8fd46a",
         "GUID:679888bc3e9c44c19ab0d62ff8701dcf",
         "GUID:9c34b9a25e1464dec91601291eade23f",
