+++ Packages/src/Editor/FirstPartyTools/RunTests/AssemblyInfo.cs
@@ -1,6 +1,7 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("UnityCLILoop.FirstPartyTools.Editor")]
+[assembly: InternalsVisibleTo("UnityCLILoop.FirstPartyTools.RunTests.TestFramework.Editor")]
 [assembly: InternalsVisibleTo("uLoopMCP.Tests.Editor")]
 [assembly: InternalsVisibleTo("uLoopMCP.Tests.PlayMode")]
 [assembly: InternalsVisibleTo("uLoopMCP.Tests.Demo.Editor")]
