+++ Packages/src/Editor/FirstPartyTools/RunTests/TestFramework.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0d99bb3d8324040458802cc0598ebe3d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
